@@ -4,7 +4,7 @@ version: 0.1.0
 homepage: https://github.com/Parassharmaa/usage_stats
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.10.0"
 
 dependencies:
