@@ -8,20 +8,18 @@ class EventInfo {
       this.count});
 
   /// Construct class from the json map
-  factory EventInfo.fromMap(Map map) => map == null
-      ? null
-      : EventInfo(
-          firstTimeStamp: map['firstTimeStamp'],
-          lastTimeStamp: map['lastTimeStamp'],
-          totalTime: map['totalTime'],
-          lastEventTime: map['lastEventTime'],
-          eventType: map['eventType'],
-          count: map['count']);
+  factory EventInfo.fromMap(Map map) => EventInfo(
+      firstTimeStamp: map['firstTimeStamp'],
+      lastTimeStamp: map['lastTimeStamp'],
+      totalTime: map['totalTime'],
+      lastEventTime: map['lastEventTime'],
+      eventType: map['eventType'],
+      count: map['count']);
 
-  final String firstTimeStamp;
-  final String lastTimeStamp;
-  final String totalTime;
-  final String lastEventTime;
-  final String eventType;
-  final String count;
+  final String? firstTimeStamp;
+  final String? lastTimeStamp;
+  final String? totalTime;
+  final String? lastEventTime;
+  final String? eventType;
+  final String? count;
 }
