@@ -6,15 +6,12 @@ class NetworkInfo {
   });
 
   /// Construct class from the json map
-  factory NetworkInfo.fromMap(Map map) => map == null
-      ? null
-      : NetworkInfo(
-          packageName: map['packageName'],
-          rxTotalBytes: map['rxTotalBytes'],
-          txTotalBytes: map['txTotalBytes']
-        );
+  factory NetworkInfo.fromMap(Map map) => NetworkInfo(
+      packageName: map['packageName'],
+      rxTotalBytes: map['rxTotalBytes'],
+      txTotalBytes: map['txTotalBytes']);
 
-  final String packageName;
-  final String rxTotalBytes;
-  final String txTotalBytes;
+  final String? packageName;
+  final String? rxTotalBytes;
+  final String? txTotalBytes;
 }
