@@ -7,19 +7,17 @@ class UsageInfo {
       this.totalTimeInForeground});
 
   /// Construct class from the json map
-  factory UsageInfo.fromMap(Map map) => map == null
-      ? null
-      : UsageInfo(
-          firstTimeStamp: map['firstTimeStamp'],
-          lastTimeStamp: map['lastTimeStamp'],
-          lastTimeUsed: map['lastTimeUsed'],
-          totalTimeInForeground: map['totalTimeInForeground'],
-          packageName: map['packageName'],
-        );
+  factory UsageInfo.fromMap(Map map) => UsageInfo(
+        firstTimeStamp: map['firstTimeStamp'],
+        lastTimeStamp: map['lastTimeStamp'],
+        lastTimeUsed: map['lastTimeUsed'],
+        totalTimeInForeground: map['totalTimeInForeground'],
+        packageName: map['packageName'],
+      );
 
-  final String packageName;
-  final String firstTimeStamp;
-  final String lastTimeStamp;
-  final String lastTimeUsed;
-  final String totalTimeInForeground;
+  final String? packageName;
+  final String? firstTimeStamp;
+  final String? lastTimeStamp;
+  final String? lastTimeUsed;
+  final String? totalTimeInForeground;
 }
