@@ -16,8 +16,8 @@ export 'src/event_usage_info.dart';
 class UsageStats {
   static const MethodChannel _channel = const MethodChannel('usage_stats');
 
-  static Future<bool> checkUsagePermission() async {
-    bool isPermission = await _channel.invokeMethod('isUsagePermission');
+  static Future<bool?> checkUsagePermission() async {
+    bool? isPermission = await _channel.invokeMethod('isUsagePermission');
     return isPermission;
   }
 
@@ -30,7 +30,7 @@ class UsageStats {
     int end = endDate.millisecondsSinceEpoch;
     int start = startDate.millisecondsSinceEpoch;
     Map<String, int> interval = {'start': start, 'end': end};
-    List events = await _channel.invokeMethod('queryEvents', interval);
+    List events = await (_channel.invokeMethod('queryEvents', interval) as FutureOr<List<dynamic>>);
     List<EventUsageInfo> result =
         events.map((item) => EventUsageInfo.fromMap(item)).toList();
     return result;
@@ -41,7 +41,7 @@ class UsageStats {
     int end = endDate.millisecondsSinceEpoch;
     int start = startDate.millisecondsSinceEpoch;
     Map<String, int> interval = {'start': start, 'end': end};
-    List configs = await _channel.invokeMethod('queryConfiguration', interval);
+    List configs = await (_channel.invokeMethod('queryConfiguration', interval) as FutureOr<List<dynamic>>);
     List<ConfigurationInfo> result =
         configs.map((item) => ConfigurationInfo.fromMap(item)).toList();
     return result;
@@ -52,7 +52,7 @@ class UsageStats {
     int end = endDate.millisecondsSinceEpoch;
     int start = startDate.millisecondsSinceEpoch;
     Map<String, int> interval = {'start': start, 'end': end};
-    List eventsStats = await _channel.invokeMethod('queryEventStats', interval);
+    List eventsStats = await (_channel.invokeMethod('queryEventStats', interval) as FutureOr<List<dynamic>>);
     List<EventInfo> result =
         eventsStats.map((item) => EventInfo.fromMap(item)).toList();
     return result;
@@ -63,7 +63,7 @@ class UsageStats {
     int end = endDate.millisecondsSinceEpoch;
     int start = startDate.millisecondsSinceEpoch;
     Map<String, int> interval = {'start': start, 'end': end};
-    List usageStats = await _channel.invokeMethod('queryUsageStats', interval);
+    List usageStats = await (_channel.invokeMethod('queryUsageStats', interval) as FutureOr<List<dynamic>>);
     List<UsageInfo> result =
         usageStats.map((item) => UsageInfo.fromMap(item)).toList();
     return result;
@@ -75,7 +75,7 @@ class UsageStats {
     int start = startDate.millisecondsSinceEpoch;
     Map<String, int> interval = {'start': start, 'end': end};
     Map usageAggStats =
-        await _channel.invokeMethod('queryAndAggregateUsageStats', interval);
+        await (_channel.invokeMethod('queryAndAggregateUsageStats', interval) as FutureOr<Map<dynamic, dynamic>>);
 //    usageAggStats.
     Map<String, UsageInfo> result = usageAggStats
         .map((key, value) => MapEntry(key as String, UsageInfo.fromMap(value)));
@@ -87,7 +87,7 @@ class UsageStats {
     int end = endDate.millisecondsSinceEpoch;
     int start = startDate.millisecondsSinceEpoch;
     Map<String, int> interval = {'start': start, 'end': end};
-    List events = await _channel.invokeMethod('queryNetworkUsageStats', interval);
+    List events = await (_channel.invokeMethod('queryNetworkUsageStats', interval) as FutureOr<List<dynamic>>);
     List<NetworkInfo> result =
         events.map((item) => NetworkInfo.fromMap(item)).toList();
     return result;
